backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-openai
openai
PyPDF2
faiss-cpu
langchain
langchain-community